Rename dynamic-module to push-messages
