+++ src/VirtoCommerce.PushMessages.Web/App/src/modules/push-messages/components/widgets/recipients-widget.vue
