+++ src/VirtoCommerce.PushMessages.Web/App/src/modules/push-messages/pages/messageDetails.ts
